Updated Functions
Moved logging config from dedicated file to within functions.py. Debugged environment variables and credentials for automated emails.
