--- piastri/functions/logging_config.py
@@ -1,24 +1,0 @@
-import logging
-
-
-def setup_logger(name: str) -> logging.Logger:
-    """
-    Set up a logger with the specified name.
-
-    Args:
-        name (str): The name of the logger.
-
-    Returns:
-        logging.Logger: The configured logger object.
-    """
-
-    logger = logging.getLogger(name)
-    handler = logging.StreamHandler()
-    formatter = logging.Formatter(
-        "%(asctime)s - %(name)s - %(levelname)s - %(message)s"
-    )
-    handler.setFormatter(formatter)
-    logger.addHandler(handler)
-    logger.setLevel(logging.INFO)
-
-    return logger
